+++ sso/user/tests/test_admin.py
@@ -27,7 +27,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
     header = (
         'created,date_joined,email,first_name,hashed_uuid,'
         'id,is_active,is_staff,is_superuser,'
-        'last_login,last_name,modified,oauth2_provider_application,page_views,user_profile,'
+        'last_login,last_name,lessoncompleted,modified,oauth2_provider_application,page_views,user_profile,'
         'utm,verification_code'
     )
 
@@ -56,7 +56,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
         row_one = (
             "{created},{date_joined},admin@example.com,,{hashed_uuid},"
             "{id},True,True,True,"
-            "{last_login},,{modified},,,,{utm},"
+            "{last_login},,,{modified},,,,{utm},"
         ).format(
             created=user.created,
             date_joined=user.date_joined,
@@ -91,7 +91,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
         row_one = (
             '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},,,{modified},,,{user_profile},''{utm},'
+            '{is_superuser},,,,{modified},,,{user_profile},''{utm},'
             '{verification_code}'
         ).format(
             created=user_one.created,
@@ -112,7 +112,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
         row_two = (
             '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},,,{modified},,,{user_profile},{utm},'
+            '{is_superuser},,,,{modified},,,{user_profile},{utm},'
             '{verification_code}'
         ).format(
             created=user_two.created,
@@ -133,7 +133,7 @@ class DownloadCaseStudyCSVTestCase(TestCase):
         row_three = (
             '{created},{date_joined},{email},,{hashed_uuid},'
             '{id},{is_active},{is_staff},'
-            '{is_superuser},{last_login},,{modified},,,{user_profile},{utm},'
+            '{is_superuser},{last_login},,,{modified},,,{user_profile},{utm},'
             '{verification_code}'
         ).format(
             created=user_three.created,
@@ -212,6 +212,7 @@ def test_download_csv_exops_not_fab(
         ('is_superuser', user_one.is_superuser),
         ('last_login', user_one.last_login),
         ('last_name', ''),
+        ('lessoncompleted', ''),
         ('modified', user_one.modified),
         ('oauth2_provider_application', ''),
         ('page_views', ''),